*test: get rid of compilation warning

+++ test/test.cpp
@@ -37,7 +37,7 @@ int main(int argc, char **argv)
   log.close_group();
   log.msg("-----------------------------------------------------------");
 
-  int total_time = log.get_total_time();
+  int total_time = (int)log.get_total_time();
   log.msg("Total time: %i:%02i", total_time / 60, total_time % 60);
   if (log.get_total_errors())
     log.msg("There are %i errors!\n", log.get_total_errors());
